+++ src/transform_fns/sha256.rs
@@ -1,11 +1,12 @@
-use sha2::{digest::generic_array::GenericArray, Digest, Sha256};
+use sha2::digest::Output;
+use sha2::{Digest, Sha256};
 
 use crate::transform_fns::TransformFn;
 
 /// SHA256-Hashing
-pub static SHA256_HASHING: TransformFn<Sha256Hash> = sha256;
+pub const SHA256_HASHING: TransformFn<Sha256Hash> = sha256;
 
-pub type Sha256Hash = GenericArray<u8, <sha2::Sha256 as Digest>::OutputSize>;
+pub type Sha256Hash = Output<Sha256>;
 
 fn sha256(input: &str) -> Sha256Hash {
     let mut hasher = Sha256::new();