+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+# v3.0.1 (2022-01-27)
+- massive performance improvements (20-40%)
+- internal code quality improved
+
 # v3.0.0 (2021-10-19)
 
 - started CHANGELOG file
 - several changes to public interface
-  - all hashing algorithms now use a common trait and are not longer
+  - all hashing algorithms now use a common trait and are no longer
     copied to a string
 - multiple internal improvements
 - better code quality